Extracted tip modal to a new component, wired it up with Redux and added new hook for Online Status
