+++ src/pages/User/index.js
@@ -1,10 +1,7 @@
 import React, { useEffect, useState, useCallback, Suspense } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { useParams } from "react-router-dom";
-import QRCode from "react-qr-code";
-import Moment from "moment";
-import CopyClipboard from "react-copy-to-clipboard";
-import { isIOS } from "react-device-detect";
+import Tooltip from "react-tooltip";
 
 import {
   getUserWall,
@@ -16,50 +13,35 @@ import {
   getUserHeader,
   getPinnedPost
 } from "../../actions/UserActions";
-import { generateGunPair } from "../../actions/AuthActions";
-import { payUser, resetPaymentRequest } from "../../actions/TransactionActions";
+import { openModal } from "../../actions/TipActions";
+import useOnlineStatus from "../../hooks/useOnlineStatus";
 import { listenPath, gunUser } from "../../utils/Gun";
-import { webTorrentClient, attachMedia } from "../../utils/Torrents";
+import { attachMedia } from "../../utils/Torrents";
 
 import Loader from "../../common/Loader";
 import Divider from "../../common/Divider";
+import TipModal from "../../common/TipModal";
 
 // Assets
 import defaultBanner from "../../images/banner-bg.jpg";
 import av1 from "../../images/av1.jpg";
 import shockLogo from "../../images/lightning-logo.svg";
 import "./css/index.css";
-import Tooltip from "react-tooltip";
 
 const Post = React.lazy(() => import("../../common/Post"));
 const SharedPost = React.lazy(() => import("../../common/Post/SharedPost"));
 
-const ONLINE_INTERVAL = 1 * 30 * 1000;
-
 const UserPage = () => {
   const dispatch = useDispatch();
   const params = useParams();
+  const { userId: publicKey } = params;
   const wall = useSelector(({ user }) => user.wall);
   const profile = useSelector(({ user }) => user.profile);
-  const me = useSelector(({ auth }) => auth.pair);
-  const paymentRequest = useSelector(
-    ({ transaction }) => transaction.paymentRequest
-  );
+  const { isOnlineApp, isOnlineNode } = useOnlineStatus(publicKey);
   // Reserved for future use
   // eslint-disable-next-line no-unused-vars
   const [userLoading, setUserLoading] = useState(true);
   const [wallLoading, setWallLoading] = useState(true);
-  const [tipModalOpen, setTipModalOpen] = useState(false);
-  const [tipLoading, setTipLoading] = useState(false);
-  const [tipAmount, setTipAmount] = useState(10);
-  const [isOnlineApp, setIsOnlineApp] = useState(false);
-  const [isOnlineNode, setIsOnlineNode] = useState(false);
-  const [copied, setCopied] = useState(false);
-
-  const [onlineCheckTimer, setOnlineCheckTimer] = useState(null);
-  const [tipMetadata, setTipMetadata] = useState({ targetType: "user" });
-
-  const publicKey = params.userId;
 
   const fetchUserData = useCallback(async () => {
     try {
@@ -91,41 +73,13 @@ const UserPage = () => {
     }
   }, [dispatch, publicKey]);
 
-  const sendTip = useCallback(async () => {
-    try {
-      setTipLoading(true);
-      await dispatch(
-        payUser({
-          senderPair: me,
-          recipientPublicKey: publicKey,
-          amount: tipAmount,
-          metadata: tipMetadata
-        })
-      );
-      setTipLoading(false);
-    } catch (err) {
-      console.error(err);
-    }
-  }, [dispatch, me, publicKey, tipAmount, tipMetadata]);
-
-  const openTipModal = (metadata = { targetType: "spontaneousPayment" }) => {
-    setTipMetadata(metadata);
-    setTipModalOpen(true);
-  };
-
-  const closeTipModal = useCallback(() => {
-    setTipModalOpen(false);
-    setTipMetadata({});
-    if (paymentRequest) {
-      dispatch(resetPaymentRequest());
-      setTipLoading(false);
-    }
-  }, [dispatch, paymentRequest]);
+  const openTipModal = useCallback(() => {
+    dispatch(openModal({ targetType: "spontaneousPayment" }));
+  }, [dispatch]);
 
   const initializeUserWall = useCallback(async () => {
     const { postId, userId, type } = params;
     await fetchUserData();
-    await dispatch(generateGunPair());
     fetchUserWall();
 
     dispatch(
@@ -141,22 +95,6 @@ const UserPage = () => {
     initializeUserWall();
 
     // Subscribe for updates
-    const lastSeenAppListener = listenPath({
-      path: "Profile/lastSeenApp",
-      gunPointer: gunUser(publicKey),
-      callback: event => {
-        dispatch(updateUserProfile({ lastSeenApp: event }));
-      }
-    });
-
-    const lastSeenNodeListener = listenPath({
-      path: "Profile/lastSeenNode",
-      gunPointer: gunUser(publicKey),
-      callback: event => {
-        dispatch(updateUserProfile({ lastSeenNode: event }));
-      }
-    });
-
     const displayNameListener = listenPath({
       path: "Profile/displayName",
       gunPointer: gunUser(publicKey),
@@ -174,8 +112,6 @@ const UserPage = () => {
     });
 
     return () => {
-      lastSeenAppListener.off();
-      lastSeenNodeListener.off();
       displayNameListener.off();
       bioListener.off();
     };
@@ -187,52 +123,6 @@ const UserPage = () => {
       false
     );
   }, [wall.posts]);
-
-  useEffect(() => {
-    if (onlineCheckTimer) {
-      clearTimeout(onlineCheckTimer);
-    }
-
-    const timer = setTimeout(() => {
-      const onlineThreshold = Moment.utc().subtract(ONLINE_INTERVAL, "ms");
-      const isOnlineNode = profile.lastSeenNode
-        ? Moment.utc(profile.lastSeenNode).isSameOrAfter(onlineThreshold)
-        : false;
-      const isOnlineApp = profile.lastSeenApp
-        ? Moment.utc(profile.lastSeenApp).isSameOrAfter(onlineThreshold)
-        : false;
-
-      setIsOnlineNode(isOnlineNode);
-      setIsOnlineApp(isOnlineApp);
-    }, ONLINE_INTERVAL);
-
-    setOnlineCheckTimer(timer);
-
-    return () => clearTimeout(onlineCheckTimer);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [profile]);
-
-  useEffect(() => {
-    const onlineThreshold = Moment.utc().subtract(1, "minutes");
-    const isOnlineNode = profile?.lastSeenNode
-      ? Moment.utc(profile?.lastSeenNode).isSameOrAfter(onlineThreshold)
-      : false;
-    const isOnlineApp = profile?.lastSeenApp
-      ? Moment.utc(profile?.lastSeenApp).isSameOrAfter(onlineThreshold)
-      : false;
-
-    setIsOnlineNode(isOnlineNode);
-    setIsOnlineApp(isOnlineApp);
-  }, [profile]);
-
-  const setCopiedStatus = useCallback(() => {
-    setCopied(true);
-    Tooltip.rebuild();
-    setTimeout(() => {
-      setCopied(false);
-      Tooltip.rebuild();
-    }, 500);
-  }, []);
 
   const username = profile.displayName ?? profile.alias;
 
@@ -265,7 +155,6 @@ const UserPage = () => {
               sharerPublicKey={publicKey}
               sharerUsername={username}
               isOnlineNode={isOnlineNode}
-              openTipModal={openTipModal}
               pinned={post.pinned}
             />
           </Suspense>
@@ -288,8 +177,6 @@ const UserPage = () => {
               username={username}
               avatar={avatar}
               publicKey={publicKey}
-              openTipModal={openTipModal}
-              webTorrentClient={webTorrentClient}
               page={post.page}
               id={post.id}
               tipValue={post.tipValue ?? 0}
@@ -359,7 +246,7 @@ const UserPage = () => {
 
           <div
             className="send-tip-btn"
-            onClick={() => openTipModal()}
+            onClick={openTipModal}
             style={{
               opacity: isOnlineNode ? 1 : 0.5,
               cursor: isOnlineNode ? "pointer" : "default"
@@ -387,79 +274,7 @@ const UserPage = () => {
       {wallLoading ? (
         <Loader text={`Loading ${wall.page >= 0 ? "More" : "Wall"} Posts...`} />
       ) : null}
-      {tipModalOpen ? (
-        <div className="tip-modal-container">
-          <div className="tip-modal-overlay" onClick={closeTipModal}></div>
-          <div className="tip-modal">
-            {tipLoading ? (
-              <div className="tip-modal-loading">
-                <Loader text="Submitting Tip Request..." />
-              </div>
-            ) : null}
-            <div className="tip-modal-head">
-              <div className="tip-modal-title">Send Tip</div>
-            </div>
-            {paymentRequest ? (
-              <div className="tip-modal-content">
-                <p className="tip-modal-instructions">
-                  We've successfully generated an invoice for you to tip, please
-                  scan the QR Code below using a Lightning wallet to pay it!
-                </p>
-                <div className="tip-modal-qr-code-container">
-                  <QRCode
-                    className="tip-modal-qr-code"
-                    value={paymentRequest}
-                    size={210}
-                    bgColor="#4db1ff"
-                    fgColor="#1b2129"
-                  />
-                </div>
-                <div className="tip-modal-action-btns">
-                  <a
-                    href={`lightning:${paymentRequest}`}
-                    className="tip-modal-action-btn"
-                  >
-                    PAY INVOICE
-                  </a>
-                  <CopyClipboard text={paymentRequest} onCopy={setCopiedStatus}>
-                    <div className="tip-modal-action-btn">
-                      {copied ? "INVOICE COPIED!" : "COPY INVOICE"}
-                    </div>
-                  </CopyClipboard>
-                </div>
-              </div>
-            ) : (
-              <div className="tip-modal-content">
-                <p className="tip-modal-instructions">
-                  Please specify the amount of sats you'd like to tip this user
-                  with below and we'll generate an invoice for you to scan.
-                </p>
-                <input
-                  className="tip-modal-input"
-                  value={tipAmount}
-                  onChange={e => setTipAmount(e.target.value)}
-                />
-              </div>
-            )}
-            {!paymentRequest ? (
-              <div className="tip-modal-footer" onClick={sendTip}>
-                <div className="tip-modal-submit">SEND TIP</div>
-              </div>
-            ) : null}
-          </div>
-        </div>
-      ) : null}
-      {!isIOS ? (
-        <a
-          href="https://wallet.shock.pub"
-          target="_blank"
-          rel="noreferrer noopener"
-          className="download-shockwallet-btn"
-        >
-          <i className="download-shockwallet-btn-icon fas fa-user"></i>
-          <p className="download-shockwallet-btn-text">Create a Shockwallet</p>
-        </a>
-      ) : null}
+      <TipModal publicKey={publicKey} />
     </div>
   );
 };
